clean solution

+++ Market Store/GoldDiscountCard.cs
@@ -5,7 +5,7 @@ namespace MarketStore
     class GoldDiscountCard : DiscountCard, IDiscountCard
     {
         private const decimal InitialDiscountRate = 0.02m;
-        private const int DiscountRateGrow = 100;
+        private const decimal DiscountRateGrow = 100;
         private const decimal MaxDiscountRate = 0.1m;
         public GoldDiscountCard(Person cardHolder) 
             : base(cardHolder)
